backend/requirements.txt
fastapi
uvicorn
pymongo
pydantic
python-dotenv

google-cloud-documentai
google-generativeai

chromadb
sentence-transformers
python-multipart

pandas
numpy
joblib
lightgbm
scikit-learn
